Refactored the getAsset function of the catalog and minor fixes
